Totally bizarre... tagalong.bin.gz disappears from resources
Update:
It happens that there is an *undocumented* culture that uses 'bin' as its name - the Nigerian dialect 'Edo'.  So the .bin.gz resource was put into a satellite assembly!

Even though it's in the project and no build errors are reported, the embedded resource 'tagalong.bin.gz' would simply disappear from the resources of the produced binary. Just ... not there at all! Renaming the file to _bin.gz or as done here, .bbin.gz - and the resource is fine.  It would not show up in manifest resource names when .bin.gz was used. FREAKY.

Workaround - use this wonky name and inject a logical name into the project.

Not sure if this is a quirk of Visual Studio 2012 or some other more esoteric rule. I was unable to find any documentation about restrictions of resource file names.  Perhaps it's buried in the behavior of 
System.Resources.Tools.StronglyTypedResourceBuilder.VerifyResourceName() or System.CodeDom.Compiler.ICodeGenerator.CreateValidIdentifier()

If you have 'foo.bin' - it works fine. But 'foo.bin.xyz' just ... disappears. Weird, weird, weird.