Add GZipMemberEntry
The goal of GZipMemberEntry is to provide a simple guess at the content of a GZIP file based on the file name and potentially the metadata in the GZIP header. It also has a naive attempt to identify concatenated GZIP files, which is allowed by the spec.  However, it does NOT actually inflate the contents -- so if the 'magic' byte pattern appears in a DEFLATEd block, the scheme falls apart.  Such files are expected to be rare.