A ROM format metadata test failed in release builds
Always had. It was testing a condition that arose specifically only in debug builds. Generally, the debug builds will throw when an error arises while parsing metadata - release builds will not. This test was expecting an exception simply because it was written and run in Debug builds.

In release, it now tests the expected result.