Add SharpZipLib implementation of GZIP access
Pretty smooth, aside from the different behavior w.r.t. concatenated (multi-member) archives. Not sure I care for the behavior here, in which ALL members are written out to one giant stream.  I suppose _technically_ it's doing the same thing that the command line tools do (which is perhaps the best kind of correct)... But then it makes a simple 'iterate until all extracted' implementation a little trickier.

Not to fear ... the simple implementation for getting GZIP member "metadata" is sufficient for the goals of this code, though it's not been tested with corrupted GZIPs nor entries with extra gunk at the end.