Small adjustments so code coverage analysis is better
Default behavior of code coverage for Boolean logic wants coverage for all possible cases in combinatorics, so testing something like `if (a && b)` will not be considered complete until all four possible combinations are tested.

These functions use the first check of a `!string.IsNullOrEmpty(s)` specifically to short-circuit so the "lack" of code coverage is bogus.  This admittedly annoying reworking produces complete code coverage for testing purposes and produces the same behavior.  It just takes more lines in the source file. :eyeroll: