Update FIleMemo to use StorageLocation
This is a big core change and will have wide-reaching ramifications. Main one is that memos no longer have their own storage access member - which was a Bad Idea.  Now, they'll build up a cache such that if the "same" path is accessed via different IStorageAccess implementations, they will have separate cache entries.

Tests kind of cheat on this, as they have a private storage for purposes of isolation from other tests.

The behavior of the memos is indirectly tested with StorageLocation in other tests.