Fix tests and code coverage
Changes to StorageLocation caused a test to fail because it expected ArgumentNullException - which no longer happens due to better safety of the Exists() method on StorageLocation.

Other change is to get code coverage of the invalid storage access' Exists method, which as above is now protected by a null check in StorageLocation. Ideally, directly using IStorageAccess can be minimized and instead always go through StorageLocation.