Make IApplicationInfo more robust
This is part of #325 to pull app info out and instead of it being an MEF import, be explicitly passed to the application at launch.

Abstract base class is added to ease implementation. This breaks the INTV.LtoFlash build (and hence overall).

Subsequent changes are needed to `SingleInstanceApplication` and elsewhere.