Update `StorageAccess.IsLocationAContainer()`
Implementations are now a little bit more lenient, in that they will always consider
a path ending with / or \ to be a container. Otherwise, they will do a check that
actually verifies existence as a directory on disk or check the entry within the archive
to determine if the path is to a directory or not.