Changes to get ROM discovery in archives closer
to working.  It's not seamless, and requires a hack in ZipArchive
to avoid crashing.  Changes need to happen for compressed archive
access to integrate cleanly - mainly to do with Stream access for ROM
identification.

These changes will successfuly delve during an 'add' operation,
but subsequent ROM identification will still fail.