Add memo for compressed archive entries
The aim with this memo implementation is to make the `IStorageAccess.Open(string)` implementation on `CompressedArchive` behave more seamlessly for callers.  The core of it is to extract an entry to a temporary location so the caller an access the entry as a normal file, in which the returned `Stream` offers normal seek support.

This is not yet tested on adding an archive with nested contents.