Normalize separators when checking for .cfg file
This is just going to tiptoe around a problem... The code really, _really_ wants to copy a .cfg file next to a .bin file. That won't work when discovering ROMs in archives, since it's somewhat rude to inject new files into the archive. That will be addressed at some point in the future.