Fix up compressed archive storage access open() method
The `IStorageAccess.Open(string)` method on compressed archive was not dealing with nested archives as desired.

Specifically, the goal of the implementation here using memos was to extract-as-needed when an entry's stream did not support seeking, as that's always used during ROM identification.

As it happens, nested archives were not handed correctly -- though at this time, the only archive format that has trouble seeking is the ZIP archive.  Hand testing of nested archives in the product revealed the shortcoming.

Now, if the entry's stream does not support seeking, and the archive is nested, find the appropriate quasi-root-location to use for memos and proceed as expected.

While fixing this up, the number of cases in which extraction occurs may have shrunk slightly, as code coverage for the code (briefly) exposed a lack of auto-testing. That has been remedied.