Update ROM identification to use StorageAccess internally
The publicly facing `IdentifyRomFiles()` helper still accepts strings, as it operates on file system paths via drag / drop and file dialog selection. The implementation beneath it immediately switches over to `StorageLocation` now, which enables dragging of archive files containing nested entries to work.

Along the way, refactored progress updating code to a simple helper method so it's not being bulky and in the way of readability of the implementation.

The larger changes are to `GetPotentialRomFiles()` which not consistently uses `StorageLocation`, embracing the `StorageLocation.ContainerExists()` rather than `Directory.Exists()` method, as well as using the updated `StorageLocationExtensions.EnumerateFiles()` method that supports recursion.