Fix a bug in ProgramFileKindHelpers
When checking file extensions, it was not passing filename to the INTV.Core file extension checking methods. Because of this, a path that had directory entries containing a period, followed by a file name with no extension, would not be identified as ROMs.

Yeah, we're talking the tip of the corner here... this was found in a .gz file that had concatenated several entries together, but could be a legitimate directory.foo/bar where bar is a valid ROM file.