Support write chunks for upload data to RAM
[Mac] Wire through the write chunk size from preferences to the UpdateDataBockToRam command

Thus far, problems surround uploading menu layouts to new carts on Mac. Such operations work by coalescing as many ROMs as possible into the onboard RAM and minimizing (ideally) the number of transfers between the host computer and the cartridge. That is, fill up as much of the 1MB of RAM as possible in "huge" single transfers, as few times as possible, to transfer all the ROMs. quickly.

It is suspected that these large 2000000 bps transfers are unreliable on macOS for <whatever reasons>. Similar problems with _READ_ operations appeared in the macOS 10.13 timeframe.

Other smaller operations appear to be working correctly, such as "Download and Play" and setting cartridge name and owner. Also of note: at least one customer has been able to successfully back up a well-loaded cartridge without error - using default chunk size for read of 512 on newer macOS. (The cart was populated on a Windows machine.)

The upshot here is that, rather than affecting ALL cartridge commands, at present only the UploadDataBlockToRam command is being adjusted -- similar to how only its counterpart was changed before. This is done by splitting out the data transfer to an overridable SendCommandPayload method. The error reports indicate failure in this command, so it gets the special sauce.

While doing this, removed an unused local that read stream size. Hope reading stream size doesn't have side effects! lol

And for good measure, I pumped the 'Scotty Factor' (a timeout) up by another factor of two when sending payload. It's a bit ridiculous, but hey, it's already nutty enough that serial ports are so brittle here anyway.

Finally: There's a bit of a deficiency perhaps in the serial logging in here... the log isn't capturing information about payload. Not sure if I left that out on purpose - but since that would inflate the log even more, potentially at the cost of introducing instability (hiccups writing to disk log file), it's going to stay out for now.