Refactor error database parsing
The YAML-specific parser is now relocated to separate partial class implementation.

The larger change is to prepare to support other formats for the error database, and to allow multi-version database formats.

Initially the plan had been tor a single error_db.yaml to suffice for all firmware versions. There were problems producing this that were not essential to product release so the databases were produced one-per-firmware release. This problem is being addressed, so multi-firmware-version database creation is happening. The code was prepared for this, but only superficially. Now it handles it better.

Additionally, the refactor lays the groundwork to introduce support for other database formats. Sadly the YAML parser selected was found to be subject to the Ms-PL license, incompatible with the GPL. Thus... XML support will be added in subsequent updates.